Moved perkpicker into its own component Added a store for state management between components Renamed Wwzplanner component to planner
